Improved audio spectrum hover logic
